Clean up BNO055 and move to utilities
Our BNO055 class can no longer be considered a magicbot component.
