+++ autonomous/segments.py
@@ -6,12 +6,12 @@ from networktables import NetworkTable
 from automations.filters import RangeFilter, VisionFilter
 from automations.manipulategear import ManipulateGear
 from automations.profilefollower import ProfileFollower
-from components.bno055 import BNO055
 from components.chassis import Chassis
 from components.gears import GearAligner, GearDepositor
 from components.range_finder import RangeFinder
 from components.vision import Vision
 from components.winch import Winch
+from utilities.bno055 import BNO055
 from utilities.profilegenerator import generate_trapezoidal_trajectory
